refactor Titles class to addPageTitles function
